+++ Project/DataModels/ThemeModel.cs
@@ -1,6 +1,6 @@
 public class ThemeModel
 {
-    public int64 ID { get; set; }
+    public int ID { get; set; }
     public string ThemeName { get; set; }
 
     public ThemeModel(int id, string themeName)
